pyproject.toml
[tool.poetry]
name = "service-rent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
djangorestframework = "^3.15.2"
drf-nested-routers = "^0.94.1"
drf-yasg = {extras = ["validation"], version = "^1.21.8"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
